refactor(components): refactor card components
fix #263
