--- packages/components/src/base/Card/cardactions.tsx
@@ -1,5 +1,0 @@
-import { CardActions as MuiCardActions, type CardActionsProps } from '@mui/material';
-
-export function CardActions(props: CardActionsProps): JSX.Element {
-  return <MuiCardActions {...props} />;
-}
