+++ packages/components/src/base/Card/index.tsx
@@ -1,3 +1,1 @@
-export { Card } from './card';
-export { CardActions } from './cardactions';
-export { CardContent } from './cardcontent';
+export { Card } from './Card';
