+++ packages/components/src/base/CardActions/CardActions.tsx
@@ -1,0 +1,8 @@
+import {
+  CardActions as MuiCardActions,
+  CardActionsProps as MuiCardActionsProps
+} from '@mui/material';
+
+export function CardActions(props: MuiCardActionsProps): JSX.Element {
+  return <MuiCardActions {...props} />;
+}
