+++ packages/components/src/base/CardActions/index.tsx
@@ -1,0 +1,1 @@
+export { CardActions } from './CardActions';
