+++ packages/components/src/base/CardContent/CardContent.tsx
@@ -1,0 +1,8 @@
+import {
+  CardContent as MuiCardContent,
+  CardContentProps as MuiCardContentProps
+} from '@mui/material';
+
+export function CardContent(props: MuiCardContentProps): JSX.Element {
+  return <MuiCardContent {...props} />;
+}
