+++ packages/components/src/base/CardContent/index.tsx
@@ -1,0 +1,1 @@
+export { CardContent } from './CardContent';
