+++ packages/components/src/base/CardHeader/CardHeader.tsx
@@ -1,0 +1,5 @@
+import { CardHeader as MuiCardHeader, CardHeaderProps as MuiCardHeaderProps } from '@mui/material';
+
+export function CardHeader(props: MuiCardHeaderProps): JSX.Element {
+  return <MuiCardHeader {...props} />;
+}
