+++ packages/components/src/base/CardHeader/index.tsx
@@ -1,0 +1,1 @@
+export { CardHeader } from './CardHeader';
