+++ packages/components/src/base/CardMedia/CardMedia.tsx
@@ -1,0 +1,5 @@
+import { CardMedia as MuiCardMedia, CardMediaProps as MuiCardMediaProps } from '@mui/material';
+
+export function CardMedia(props: MuiCardMediaProps): JSX.Element {
+  return <MuiCardMedia {...props} />;
+}
