+++ packages/components/src/base/CardMedia/index.tsx
@@ -1,0 +1,1 @@
+export { CardMedia } from './CardMedia';
