+++ packages/components/src/index.tsx
@@ -5,6 +5,10 @@ export * from './base/Box';
 export * from './base/Button';
 export * from './base/ButtonGroup';
 export * from './base/Card';
+export * from './base/CardActions';
+export * from './base/CardContent';
+export * from './base/CardHeader';
+export * from './base/CardMedia';
 export * from './base/Checkbox';
 export * from './base/Chip';
 export * from './base/ClickAwayListener';
